updated build/release scripts

+++ scripts/build_conda_recipe.sh
@@ -1,8 +1,13 @@
-#!/bin/bash -uex
+#!/bin/bash -ex
 # Usage:
 #
 #    $ ./scripts/build_conda_recipe.sh v1.2.3
 #
-echo $1>__conda_version__.txt
+if [[ $1 != v* ]]; then
+    echo "Argument does not start with 'v'"
+    exit 1
+fi
+./scripts/check_clean_repo_on_master.sh
+echo ${1#v}>__conda_version__.txt
 trap "rm __conda_version__.txt" EXIT SIGINT SIGTERM
 conda build conda-recipe
